+++ plugin_ide.ui/src-lang/melnorme/util/swt/components/fields/CheckBoxField.java
@@ -68,6 +68,7 @@ public class CheckBoxField extends LabelledFieldComponent<Boolean> {
 		checkBox.setSelection(getFieldValue());
 	}
 	
+	@Override
 	public void setEnabled(boolean enabled) {
 		SWTUtil.setEnabledIfOk(checkBox, enabled);
 	}