+++ plugin_ide.ui/src-lang/melnorme/util/swt/components/fields/ComboOptionsField.java
@@ -140,6 +140,7 @@ public class ComboOptionsField extends LabelledFieldComponent<String> {
 		}
 	}
 	
+	@Override
 	public void setEnabled(boolean enabled) {
 		SWTUtil.setEnabledIfOk(getFieldControl(), enabled);
 	}